structure the utils folder
