+++ editor/core/utils/index.ts
@@ -3,10 +3,16 @@ import { reactiveState } from "core/state";
 
 import { notification as _Notification } from "./Notification";
 
+/**
+ * Get [posMap, posRecord] from posAgent
+ */
 export async function getPos() {
   return await Promise.all([posAgent.getPosMap(), posAgent.getPosRecord()]);
 }
 
+/**
+ * Get [controlMap, controlRecord] from controlAgent
+ */
 export async function getControl() {
   return await Promise.all([
     controlAgent.getControlMap(),
@@ -14,6 +20,9 @@ export async function getControl() {
   ]);
 }
 
+/**
+ * Get ledMap from ledAgent
+ */
 export async function getLedMap() {
   return await ledAgent.getLedMap();
 }
@@ -32,3 +41,6 @@ export const notification = _Notification;
 export * from "./math";
 export * from "./localStorage";
 export * from "./color";
+export * from "./fade";
+export * from "./frame";
+export * from "./led";
